pyproject.toml
[tool.poetry]
name = "chipa-text-to-vector"
version = "2.0.0"
description = "Text-to-vector service: grpc API for spacy nlp to transform text to vector"
license = "MIT"
authors = ["Chernetsov Artyom <archertech.ru>"]
readme = "README.md"
homepage = "https://gitlab.com/autofaqer-group/nlp-text-to-vector"
repository = "https://gitlab.com/autofaqer-group/nlp-text-to-vector"
#documentation = "https://docs.temporal.io/docs/python"
packages = [
    { include = "**/*.py", from = "."  }
]

[tool.poetry.urls]
"Bug Tracker" = "https://gitlab.com/autofaqer-group/nlp-text-to-vector/-/issues"

[tool.poetry.dependencies]
python = "^3.10"
spacy = "^3.5"
pydantic = "^1.10"
grpcio = "^1.51"
protobuf = "^4.22"

[tool.poetry.dev-dependencies]
grpcio-tools="^1.51"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
pytest = "^7.1.2"
pytest-asyncio = "^0.18.3"

[tool.poe.tasks]
format = [{cmd = "black ."}, {cmd = "isort ."}]
lint = [{cmd = "black --check ."}, {cmd = "isort --check-only ."}, {ref = "lint-types" }]
lint-types = "mypy --check-untyped-defs --namespace-packages ."
test = "pytest"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.mypy]
ignore_missing_imports = true
namespace_packages = true

[[tool.mypy.overrides]]
module = "aiohttp.*"
ignore_errors = true
